PLAT-301: renamed requests to handlers
