+++ ledger-core/v2/virtual/handlers/vcallrequest.go
@@ -3,7 +3,7 @@
 // This material is licensed under the Insolar License version 1.0,
 // available at https://github.com/insolar/assured-ledger/blob/master/LICENSE.md.
 
-package request
+package handlers
 
 import (
 	"github.com/insolar/assured-ledger/ledger-core/v2/conveyor/smachine"
